// web/src/shared/api/generated/model/teacherThreadContextResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface TeacherThreadContextResult {
  subjectId: string;
  subjectName: string;
  unitId: string;
  unitName: string;
  lessonId: string;
  lessonName: string;
  /** @nullable */
  questionId: string | null;
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  questionVersion: number | string | null;
  /** @nullable */
  questionStem: string | null;
  /** @nullable */
  attemptId: string | null;
}